+++ bertserini/utils.py
@@ -3,22 +3,27 @@ import string
 import logging
 import json
 import re
+import zhon
+import numpy as np
+
 
 def strip_accents(text):
     return "".join(char for char in unicodedata.normalize('NFKD', text)
                    if unicodedata.category(char) != 'Mn')
 
+
 def choose_best_answer(final_answers, score_computer,
                        paragraph_score_weight, phrase_score_weight, mode="origin"):
-
     scored_answers = get_voted_answers(final_answers, score_computer,
-                                        paragraph_score_weight, phrase_score_weight, mode)
+                                       paragraph_score_weight, phrase_score_weight, mode)
     sorted_answers = sorted(scored_answers, key=lambda k: k['total_score'], reverse=True)
 
     return sorted_answers[0]
 
+
 def weighted_score(paragraph_score, phrase_score, paragraph_weight=0.5, phrase_weight=0.5):
     return paragraph_score * paragraph_weight + phrase_score * phrase_weight
+
 
 def get_type(sent):
     ts = ['Who', 'Why', 'What', 'Which', 'When', 'How', 'Where']
@@ -31,6 +36,7 @@ def get_type(sent):
                 return tp
     return tp
 
+
 def get_voted_answers(answerlist, score_computer, paragraph_score_weight, phrase_score_weight, mode="origin"):
     if mode == "origin":
         return get_scored_answers(answerlist, score_computer, paragraph_score_weight, phrase_score_weight)
@@ -38,10 +44,10 @@ def get_voted_answers(answerlist, score_computer, paragraph_score_weight, phrase
         return get_scored_answers(answerlist[0], score_computer, paragraph_score_weight, phrase_score_weight)
     else:
         answer_dict = {}
-        #base_ans = get_scored_answers(answerlist[0], score_computer, paragraph_score_weight, phrase_score_weight)
-        #print(base_ans)
+        # base_ans = get_scored_answers(answerlist[0], score_computer, paragraph_score_weight, phrase_score_weight)
+        # print(base_ans)
         answerlist = answerlist
-        #answerlist = [answerlist[0]]
+        # answerlist = [answerlist[0]]
         answers = get_scored_answers(answerlist[0], score_computer, paragraph_score_weight, phrase_score_weight)
         for ans in answers:
             answer_text = ans['answer']
@@ -49,11 +55,11 @@ def get_voted_answers(answerlist, score_computer, paragraph_score_weight, phrase
             answer_score = ans['total_score']
             if answer_text not in answer_dict:
                 answer_dict[answer_text] = {
-                        "count": 1,
-                        "total_scores": [1 * answer_score],
-                        "sentences": [answer_sentence],
-                        "answer_text": [answer_text]
-                        }
+                    "count": 1,
+                    "total_scores": [1 * answer_score],
+                    "sentences": [answer_sentence],
+                    "answer_text": [answer_text]
+                }
             else:
                 answer_dict[answer_text]['count'] += 1
                 answer_dict[answer_text]['total_scores'].append(1 * answer_score)
@@ -74,11 +80,11 @@ def get_voted_answers(answerlist, score_computer, paragraph_score_weight, phrase
 
         return combined_answers
 
+
 def get_scored_answers(final_answers, score_computer, paragraph_score_weight, phrase_score_weight):
-    
     scored_answers = []
     for ans_id, ans in enumerate(final_answers):
-        #print(ans)
+        # print(ans)
         paragraph_score = ans['paragraph_score']
         phrase_score = ans['phrase_score']
         total_score = score_computer(paragraph_score, phrase_score, paragraph_score_weight, phrase_score_weight)
@@ -87,6 +93,7 @@ def get_scored_answers(final_answers, score_computer, paragraph_score_weight, ph
         scored_answers.append(new_answer)
         # break
     return scored_answers
+
 
 def convert_squad_to_list(squad_filename):
     data = json.load(open(squad_filename, 'r'))
@@ -101,6 +108,7 @@ def convert_squad_to_list(squad_filename):
                 answers = qa["answers"]
                 converted_data.append({"id": id_, "question": question, "answers": answers, "context": text})
     return converted_data
+
 
 def init_logger(bot):
     # create logger with 'spam_application'
@@ -122,14 +130,17 @@ def init_logger(bot):
     logger.addHandler(ch)
     return logger
 
+
 def split_title(paragraph):
     sents = paragraph.split(".")
-    text =  ".".join(sents[1:]).strip()
+    text = ".".join(sents[1:]).strip()
     title = sents[0].strip()
     return title, text
 
+
 def normalize_answer(s):
     """Lower text and remove punctuation, articles and extra whitespace."""
+
     def remove_articles(text):
         return re.sub(r'\b(a|an|the)\b', ' ', text)
 
@@ -145,6 +156,7 @@ def normalize_answer(s):
 
     return white_space_fix(remove_articles(remove_punc(lower(s))))
 
+
 def normalize_text(s):
     def remove_punc(text):
         exclude = set(string.punctuation)
@@ -155,8 +167,10 @@ def normalize_text(s):
 
     return remove_punc(lower(s))
 
+
 def normalize_chinese_text(s):
     def remove_punc(text):
         exclude = set(zhon.hanzi.punctuation)
         return ''.join(ch for ch in text if ch not in exclude)
+
     return remove_punc(s)
